requirements.txt
Django>=4.2,<5
psycopg2-binary>=2.9
django-environ>=0.11
gunicorn>=21.0
whitenoise>=6.0
cryptography>=41.0
